requirements.txt
certifi==2019.9.11
chardet==3.0.4
gnupg==2.3.1
idna==2.8
kitchen==1.2.6
Markdown==3.1.1
psutil==5.6.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
taskw==1.2.0
urllib3==1.25.7
